refactor: 로또 번호 당첨확인과 수익률 체크하는 기능를 컨트롤러에서  저장소로 분리
